The DTLS handshake is now finishing and media data is being received.
Now the DTLS handshake is completing and we start to receive media data
(rtp/rtcp) from the chrome browser client. However, some things yet to
be implemented in dtls are
- verification of peer certificate
- dtls alerts
- error handling and state maintenance

The other thing to be done is the change in the state of the peer
connection once the dtls handshake is completed successfully as well as
scenarios where the fingerprint match fails.