Adding peer cert fingerprint verification
As part of this commit, support for verifying the peer certificate
fingerprint against the one received in the sdp media description is
added. However, yet to add support for extracting the keying material
and setting up of the srtp sessions, which must be done next.

Apart from the above, also fixed the select return error in the rtc
signaling module whenever the media server client disconnected (socket).