Able to establish 2 parallel peer connections with 2 browsers
This commit includes changes where we are able to establish two parallel
peer connections with two different browsers. One browser is the
transmitter (broadcast) and the other is the receiver of media.
Establishing the media pipe line between the two is the next logical
step. This is something I am very excited about and also look at the cpu
consumption for one receiver!!!