Fixed the SRTP unprotect() error on the receiving chrome client
Fixed the srtp unprotect() error on the receiving end of the chrome
client. This error was seen when our server sent/forwarded the media to
the chrome client. When the chrome browser was run with logging enabled,
then this srtp_unprotect() error was seen with err=7.

This was because our server was using wrong srtp key and salt when
encoding the packets. This has not been fixed in this commit. However,
though the srtp_unprotect() error is not longer observed at the chrome
bwowser end, still no audio & video is being rendered. This needs to be
debugged further.

One potential cause can be the ssrc value being different in the sdp
answer and the actual ssrc in the sent packets. Needs to be probed
further.