Finally! Now being able to render audio and video on the receiver end
Finally! Achieved the first major milestone. Now chrome on the receiver
end is able to render the audio and video. However, video appears after
a delay of 30 seconds or so, while audio seems to be immediate. Need to
look into the delay of the video rendering.

The fix was related to the ssrc values being conveyed in the ssrc media
attribute in the sdp. This needs to be same as the ssrc values being
sent within the rtp header. Also another issue was that, with the latest
chrome video media contains two SSRC values and this had to be fixed in
our code.

This is the first milestone to be achieved since the start. This needs
to be celebrated.