Fixed the issue of sending RTCP packets to receivers
This took some time, but finally. The media server was forwarding the
RTCP packets to the receivers by srtp protecting like rtp packets. This
was detected when google chrome client was run in debug mode with logs
enabled, and it was seen that chrome was throwing up srtp_unprotect()
errors for RTCP packets.

Now with this commit, I am not seeing any errors on the google chrome
client logs about rtcp packets. However, video still takes a while to be
rendered which is a bigger problem and needs to be handled asap.