Successfully sent FIR request and received intra-frame video from chrome
With the changes in this commit, we are able to send a properly
formatted RTCP feedback FIR (Full Intra Request) command to chrome which
chrome understood and sent back an intra-frame for video. With this we
do not see the video being rendered after some time in the receivers. As
soon as receivers join the session, they are able to see the video
properly.

Looking into the wertc-internals stats for receiver in chrome, we see
only one PLI being sent. This solves one of our current headache issues,
so we can look to the next set of challenges.