Fixing an issue with fingerprint comparision when fp contains null char
This is a probable (?) fix for  a scenario where the fingerprint
calculated from the received cert contains a NULL character. So using
strcpy() is not the right api to use for such operations. Due to this we
have observed fp matching failing. In order to test this, we need to get
a scenario where the fp contains NULL. As of now, checking this in
without much testing.