Adding initial support for data channels using usrsctp library
Adding initial support for data channels using the usrsctp library. This
commit has a very basic support which is just initialization of the
usrsctp library. Still going through all the rfc specs and drafts for
the data channel protocol and data support. And also going through the
usrsctp library sample code and the Manual. The info is scarce about the
library usage in the context of the webrtc, hence going through the
mailing list as well.