Adding fix for incomplete DTLS handshake wth latest Openssl lib
Adding fix for incomplete DTLS handshake wth latest Openssl lib on
Ubuntu. Due to some changes in the latest openssl library, the dtls
handshake was not completing and hence the peerconnection was never
getting established. This has now been fixed in this commit, by calling
an additional openssl api -  SSL_CTX_set_read_ahead(ssl_ctx, 1);

With this additional openssl api call, things seem to be ok. DTLS
handshake completes successfully and peerconnection also gets
established successfully. The source for this fix was -
https://mta.openssl.org/pipermail/openssl-dev/2015-January/000400.html

Some other changes in this commit are

1] As mentioned above, fix for dtls handshake with latest openssl
2] Complete rewrite of dtls handshake records re-transmission. Earlier
we used to have a timer firing every x msecs, and we would re-transmit
the previously sent message. This was a temporary workaround that had
been done to make some quick progress. In this commit, that mechanism
has been removed and we make use of the timeout apis provided by Openssl
- DTLSv1_get_timeout() and DTLSv1_handle_timeout(). The source for this
  info was http://sctp.fh-muenster.de/DTLS.pdf.