Fixing a memory leak and an invalid memory read issue
This commit has 2 changes.

1] in pc module, invalid read has been fixed where we were trying to
deallocate srtp session when it was not created at all. This happens
when we try to destroy a PC session which never to ACTIVE state probably
because ice did not complete or dtls fingerprint matching failed, etc.

2] in dtls library, the sctp application data was never freed after
being passed to the application. This has been done now in this commit.