Fixing the invalid memory access when sctp association is destroyed
Earlier when we were using a graceful shutdown procedure to tear down
the SCTP association we were seeing invalid memory reads (valgrind) and
hence crash. This was because the sctp shutdown procedure involves
protocol data exchanges and hence it happens asynchromously and
application does not know when it is successfully torn down. Due to
this, application does not know when to tear down the app session
context.

Now in this commit, we are setting the SO_LINGER to 0 which triggers an
ABORT message and the association is immediately torn down. With this
change, we are not seeing any more invalid memory access in valgrind.

Still, believe that it would be better if usrsctp provides a means to
know when the associtation has been torn down completely when graceful
shutdown is used. This is currently missing usrsctp library.