Fixes related to removing of sockets from epoll loop
This commit contains fixes related to removing of socket fd of a
participant when the participant is removed from the session. Also the
socket is also being closed now after removing from epoll list. It was
being noticed that the livecast process was going into an infinite loop
of epoll_wait() and recvfrom() after some time and used to consume all
the available cpu. Hopefully these fixes are going to make sure such a
scenario does not happen again.